Settings!

+++ handlers.py
@@ -3,6 +3,7 @@ import logging
 import os
 from textwrap import dedent
 import traceback
+from typing import Optional
 from lunchable import LunchMoney, TransactionUpdateObject
 from telegram import InlineKeyboardButton, InlineKeyboardMarkup, Update
 from telegram.ext import ContextTypes
@@ -17,6 +18,7 @@ from budget_messaging import (
 )
 from lunch import NoLunchToken, get_lunch_client, get_lunch_client_for_chat_id
 from tx_messaging import send_plaid_details, send_transaction_message
+from utils import get_chat_id
 
 logger = logging.getLogger("handlers")
 
@@ -33,27 +35,43 @@ async def handle_start(update: Update):
             Only one token is supported per chat.
             """
         ),
-        parse_mode=ParseMode.MARKDOWN,
+        parse_mode=ParseMode.MARKDOWN_V2,
         disable_web_page_preview=True,
     )
 
 
-async def handle_register_token(update: Update, context: ContextTypes.DEFAULT_TYPE):
-    token = update.message.text.split(" ")[1]
+EXPECTING_TOKEN = "token"
+current_expectation = {}
+
+
+async def handle_register_token(
+    update: Update, context: ContextTypes.DEFAULT_TYPE, token_override: str = None
+):
+    # if the message is empty, ask to provide a token
+    if token_override is None and len(update.message.text.split(" ")) < 2:
+        msg = await context.bot.send_message(
+            chat_id=update.message.chat_id,
+            text="Please provide a token to register",
+        )
+        current_expectation[get_chat_id(update)] = {
+            "expectation": EXPECTING_TOKEN,
+            "msg_id": msg.message_id,
+        }
+        return
+
+    if token_override is not None:
+        token = token_override
+    else:
+        token = update.message.text.split(" ")[1]
 
     # delete the message with the token
     await context.bot.delete_message(
         chat_id=update.message.chat_id, message_id=update.message.message_id
     )
-    await context.bot.send_message(
-        chat_id=update.message.chat_id,
-        text="Deleted token for security purposes",
-    )
-
-    # make sure the token is valid
-    lunch = get_lunch_client(token)
 
     try:
+        # make sure the token is valid
+        lunch = get_lunch_client(token)
         lunch_user = lunch.get_user()
         get_db().save_token(update.message.chat_id, token)
 
@@ -65,6 +83,10 @@ async def handle_register_token(update: Update, context: ContextTypes.DEFAULT_TY
                 Hello {lunch_user.user_name}!
 
                 Your token was successfully registered. Will start polling for unreviewed transactions.
+
+                Use /settings to change my behavior.
+
+                (_I deleted token you provided for security purposes_)
                 """
             ),
             parse_mode=ParseMode.MARKDOWN,
@@ -74,13 +96,13 @@ async def handle_register_token(update: Update, context: ContextTypes.DEFAULT_TY
             chat_id=update.message.chat_id,
             text=dedent(
                 f"""
-                Failed to register token:
+                Failed to register token `{token}`:
                 ```
                 {e}
                 ```
                 """
             ),
-            parse_mode=ParseMode.MARKDOWN,
+            parse_mode=ParseMode.MARKDOWN_V2,
         )
 
 
@@ -341,19 +363,21 @@ async def handle_mark_unreviewed(update: Update, context: ContextTypes.DEFAULT_T
 async def handle_errors(update: Update, context: ContextTypes.DEFAULT_TYPE):
     """Log Errors caused by Updates."""
     if isinstance(context.error, NoLunchToken):
-        await update.message.reply_text(
+        await context.bot.send_message(
+            chat_id=get_chat_id(update),
             text=dedent(
                 """
                 No token registered for this chat. Please register a token using:
                 `/register <token>`
                 """
             ),
-            parse_mode=ParseMode.MARKDOWN,
+            parse_mode=ParseMode.MARKDOWN_V2,
         )
         return
     if os.environ.get("DEBUG"):
         error = context.error
-        await update.message.reply_text(
+        await context.bot.send_message(
+            chat_id=get_chat_id(update),
             text=dedent(
                 f"""
                 An error occurred:
@@ -362,10 +386,166 @@ async def handle_errors(update: Update, context: ContextTypes.DEFAULT_TYPE):
                 ```
                 """
             ),
-            parse_mode=ParseMode.MARKDOWN,
+            parse_mode=ParseMode.MARKDOWN_V2,
         )
     else:
         logger.error(
             f"Update {update} caused error {context.error}",
             exc_info=context.error,
+        )
+
+
+async def handle_generic_message(update: Update, context: ContextTypes.DEFAULT_TYPE):
+    # if waiting for a token, register it
+    expectation = current_expectation.get(get_chat_id(update), None)
+    if expectation["expectation"] == EXPECTING_TOKEN:
+        current_expectation[get_chat_id(update)] = None
+
+        await context.bot.delete_message(
+            chat_id=get_chat_id(update), message_id=expectation["msg_id"]
+        )
+
+        return await handle_register_token(
+            update, context, token_override=update.message.text
+        )
+
+    logger.info(f"Received unexpected message: {update.message.text}")
+
+
+def get_current_settings_text(chat_id: int) -> Optional[str]:
+    settings = get_db().get_current_settings(chat_id)
+    if settings is None:
+        return None
+
+    poll_interval = settings["poll_interval_secs"]
+    if poll_interval is None or poll_interval == 0:
+        poll_interval = "Disabled"
+    else:
+        if poll_interval < 3600:
+            poll_interval = f"`{poll_interval // 60} minutes`"
+        elif poll_interval < 86400:
+            if poll_interval // 3600 == 1:
+                poll_interval = "`1 hour`"
+            else:
+                poll_interval = f"`{poll_interval // 3600} hours`"
+        else:
+            if poll_interval // 86400 == 1:
+                poll_interval = "`1 day`"
+            else:
+                poll_interval = f"`{poll_interval // 86400} days`"
+
+    token = settings["token"]
+
+    return dedent(
+        f"""
+        *Current settings*
+
+        Poll interval: {poll_interval}
+        API token: ||{token}||
+        """
+    )
+
+
+def get_settings_buttons() -> InlineKeyboardMarkup:
+    return InlineKeyboardMarkup(
+        [
+            [
+                InlineKeyboardButton(
+                    "Change poll interval",
+                    callback_data=f"changePollInterval",
+                ),
+                InlineKeyboardButton(
+                    "Change token",
+                    callback_data=f"registerToken",
+                ),
+            ],
+            [
+                InlineKeyboardButton(
+                    "Done",
+                    callback_data="doneSettings",
+                )
+            ],
+        ]
+    )
+
+
+async def handle_settings(update: Update, context: ContextTypes.DEFAULT_TYPE):
+    """Sends a message with the current settings."""
+    settings_text = get_current_settings_text(update.message.chat_id)
+    if settings_text is None:
+        await update.message.reply_text(
+            text="No settings found for this chat. Did you register a token?",
+        )
+        return
+
+    await update.message.reply_text(
+        text=settings_text,
+        reply_markup=get_settings_buttons(),
+        parse_mode=ParseMode.MARKDOWN_V2,
+    )
+
+
+async def handle_set_token_from_button(update: Update, _: ContextTypes.DEFAULT_TYPE):
+    msg = await update.callback_query.edit_message_text(
+        text="Please provide a token to register",
+    )
+    current_expectation[get_chat_id(update)] = {
+        "expectation": EXPECTING_TOKEN,
+        "msg_id": msg.message_id,
+    }
+
+
+async def handle_change_poll_interval(
+    update: Update, context: ContextTypes.DEFAULT_TYPE
+):
+    """Changes the poll interval for the chat."""
+    if "_" in update.callback_query.data:
+        poll_interval = int(update.callback_query.data.split("_")[1])
+        get_db().update_poll_interval(get_chat_id(update), poll_interval)
+        await update.callback_query.edit_message_text(
+            text=f"_Poll interval updated_\n\n{get_current_settings_text(get_chat_id(update))}",
+            reply_markup=get_settings_buttons(),
+            parse_mode=ParseMode.MARKDOWN_V2,
+        )
+    else:
+        await update.callback_query.edit_message_text(
+            text="Please choose the new poll interval in minutes...",
+            reply_markup=InlineKeyboardMarkup(
+                [
+                    [
+                        InlineKeyboardButton(
+                            "5 minutes",
+                            callback_data=f"changePollInterval_300",
+                        ),
+                        InlineKeyboardButton(
+                            "30 minutes",
+                            callback_data=f"changePollInterval_1800",
+                        ),
+                        InlineKeyboardButton(
+                            "1 hour",
+                            callback_data=f"changePollInterval_3600",
+                        ),
+                    ],
+                    [
+                        InlineKeyboardButton(
+                            "4 hours",
+                            callback_data=f"changePollInterval_14400",
+                        ),
+                        InlineKeyboardButton(
+                            "24 hours",
+                            callback_data=f"changePollInterval_86400",
+                        ),
+                        InlineKeyboardButton(
+                            "Disable",
+                            callback_data=f"changePollInterval_0",
+                        ),
+                    ],
+                    [
+                        InlineKeyboardButton(
+                            "Cancel",
+                            callback_data="cancelPollIntervalChange",
+                        )
+                    ],
+                ]
+            ),
         )
